requirements.txt
aiohttp
certifi
discord.py
feedparser
pymongo
python-dotenv
requests
python-logging-discord-handler